refactor: feed-detail view 분리 및 공통 유틸 entities/feed로 이동
