+++ apps/web/app/(with-tab)/feed/[id]/page.tsx
@@ -1,3 +1,3 @@
-import { FeedDetailPage } from '@/views/feed/[id]';
+import { FeedDetailPage } from '@/views/feed-detail';
 
 export default FeedDetailPage;
