+++ apps/web/src/entities/feed/index.ts
@@ -1,4 +1,9 @@
-export { formatLikeCount, formatStudyDuration } from './lib/format';
+export {
+  formatLikeCount,
+  formatStudyDate,
+  formatStudyDuration,
+  formatTimeAgo,
+} from './lib/format';
 export { FEED_QUERY_KEY } from './model/query-keys';
 export type { FeedPage, FeedPost, FeedTag } from './model/types';
 export { default as FeedStatsSummary } from './ui/FeedStatsSummary';
